requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
WTForms==3.1.2
email-validator==2.1.1
qrcode==7.4.2
Pillow==10.3.0
PyMySQL==1.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
